added new tests for better covering of tuples inside with statements
