+++ tests/data/cases/context_managers_39.py
@@ -90,25 +90,6 @@ with (x, y) as z:
     pass
 
 
-# don't remove the brackets here, it changes the meaning of the code.
-# even though the code will always trigger a runtime error
-with (name_5, name_4), name_5:
-    pass
-
-
-def test_tuple_as_contextmanager():
-    from contextlib import nullcontext
-
-    try:
-        with (nullcontext(),nullcontext()),nullcontext():
-            pass
-    except TypeError: 
-        # test passed
-        pass
-    else:
-        # this should be a type error
-        assert False
-
 # output
 
 
@@ -202,23 +183,3 @@ async def func():
 # don't remove the brackets here, it changes the meaning of the code.
 with (x, y) as z:
     pass
-
-
-# don't remove the brackets here, it changes the meaning of the code.
-# even though the code will always trigger a runtime error
-with (name_5, name_4), name_5:
-    pass
-
-
-def test_tuple_as_contextmanager():
-    from contextlib import nullcontext
-
-    try:
-        with (nullcontext(), nullcontext()), nullcontext():
-            pass
-    except TypeError:
-        # test passed
-        pass
-    else:
-        # this should be a type error
-        assert False
